feat(wfmgen): 64-bit LFSR + primitive-polynomial table for lengths 2..64
Widen the PN/LFSR engine to a full 64-bit register and extend the maximal-length
primitive-polynomial table from n≤16 to n=2..64.

- pn_core: poly/seed/reg/mask → uint64_t; pn_create accepts length up to 64
  (full-width mask at 64). pn_step/generate unchanged in shape.
- synth_mls_poly() → uint64_t, table for n=2..64, generated from galois-verified
  primitive polynomials (period 2^n-1); the n=2..16 values are byte-identical to
  the previous hand-table.
- propagate uint64 pn_poly through synth_create (+ the sacred create_impl local,
  which previously truncated the 64-bit table value for n>32 — a real bug),
  the composer segment + wfmgen-1 JSON + the wfmgen CLI, and the Python bindings
  (Synth/PN parse pn_poly/poly/seed as unsigned long long). wavegen faces
  regenerated (--pn_poly → strtoull).
- test_pn_core: length-64 register (mask, high-half use, no collapse), MLS period
  for n=7/17/20 (exercises the n>32 path), and table coverage (nonzero 2..64,
  zero outside). Full suite green; drift clean.

Note: pn_poly in the wfmgen-1 JSON is a number (cJSON double), so explicit
polynomials above 2^53 should use the CLI/API; auto (pn_poly=0) is exact at any
length.

+++ native/inc/wfmgen/wfm_compose.h
@@ -53,7 +53,7 @@ typedef struct {
     uint32_t seed;     /* PRNG / LFSR seed */
     int sps;           /* samples per symbol / chip */
     int pn_length;     /* LFSR register length */
-    uint32_t pn_poly;  /* 0 → MLS poly for the length */
+    uint64_t pn_poly;  /* 0 → MLS poly for the length */
     size_t num_samples; /* on-time (samples) */
     size_t off_samples; /* off-time gap after the segment (samples) */
 } wfm_segment_t;
